Shared launch configuration.
JAXBContextResolver was never getting used because I forgot the @Provider annotation.  Once I used it, I discovered that jersey-json NATURAL requires JAXB 2.1.12+, while JDK6 on
ly has JAXB 2.1.3.  Thus, I grabbed the latest JAXB 2.2 and dropped it into the lib for the webapp.  Thankfully, the webapp looks like it uses a delegate-first classloader, so I
 don't need to worry about java.endorsed.dirs to override the JDK6 JAXB api.

Added a real session object so that I could query for users by session.  This might be overkill, and I might be able to do a contains query for a String on an object in JDO, but
 it was good practice regardless.

Reorganized .gitignores.  Now, .gitignores are relative, this should make refactoring much smoother, if necessary.