Added PushUserServlet that allows an admin user to merge multiple users into the database.  Added a resource that does nothing but guarantee that a session has been created.  This should not have security consequences, so it does not require an admin user.
Still need to figure out how to test email.  In the dev server, email just prints to the console, so I should be able to hook up a custom logger to the server, and check the log to test that the email was sent.  Once that is complete, I'll just need to implement user deletion, and users will be complete.